requirements.txt
undetected-chromedriver
setuptools
curl_cffi
pydoll-python